Derivative dictionaries, refactored cost functions
To permit recurrent networks, backpropagation needs to return the derivatives with respect to multiple different variables: the complete derivative, if the derivative has reached a seed, as well as the chain propagated down to a branch. The chain propagation is further backpropagated, and ultimately gradients along different paths are summed together.